重构与修复：缓存Pub/Sub失效、秒杀自动上下架、修逻辑bug、清理死代码、加注释和索引

+++ src/main/java/my_mall/entity/po/Goods.java
@@ -24,6 +24,6 @@ public class Goods implements Serializable {
     private   Boolean sellStatus;
     private   LocalDateTime createTime;
     private  LocalDateTime updateTime;
-    private   Integer createUser;
-    private   Integer updateUser;
+    private   Long createUser;
+    private   Long updateUser;
 }
